feat: 커스텀 AuthenticationProvider 도입 - JwtUtil과 역할 분리
